LAP10_15

+++ controllers/admin.js
@@ -1,4 +1,7 @@
 const mongoose = require('mongoose');
+
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -19,8 +22,7 @@ exports.postAddProduct = (req, res, next) => {
   const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(image)
-  if(!image){
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
@@ -55,10 +57,10 @@ exports.postAddProduct = (req, res, next) => {
     });
   }
 
-  const imageUrl =image.path
+  const imageUrl = image.path;
 
   const product = new Product({
-    // _id :new mongoose.Types.ObjectId('62afc344f22c79230c821b81'),
+    // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
     title: title,
     price: price,
     description: description,
@@ -73,7 +75,6 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
-      // console.log(errors.array());
       // return res.status(500).render('admin/edit-product', {
       //   pageTitle: 'Add Product',
       //   path: '/admin/add-product',
@@ -85,12 +86,12 @@ exports.postAddProduct = (req, res, next) => {
       //     price: price,
       //     description: description
       //   },
-      //   errorMessage: 'Database operation  failed please try again.',
+      //   errorMessage: 'Database operation failed, please try again.',
       //   validationErrors: []
       // });
-      // res.redirect('/500')
+      // res.redirect('/500');
       const error = new Error(err);
-      error.httpStatusCode =500
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -103,7 +104,6 @@ exports.getEditProduct = (req, res, next) => {
   const prodId = req.params.productId;
   Product.findById(prodId)
     .then(product => {
-  
       if (!product) {
         return res.redirect('/');
       }
@@ -117,36 +117,37 @@ exports.getEditProduct = (req, res, next) => {
         validationErrors: []
       });
     })
-    .catch(err => 
-     { const error = new Error(err);
-      error.httpStatusCode =500
-      return next(error);}
-      );
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
-      editing: false,
+      pageTitle: 'Edit Product',
+      path: '/admin/edit-product',
+      editing: true,
       hasError: true,
       product: {
-          title: title,
-          price: price,
-          description: description
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDesc,
+        _id: prodId
       },
-      errorMessage: 'Attached file is not an image.',
-      validationErrors: []
-  });
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
+    });
   }
 
   Product.findById(prodId)
@@ -157,16 +158,20 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
-    .catch(err =>
-       { const error = new Error(err);
-      error.httpStatusCode =500
-      return next(error);});
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getProducts = (req, res, next) => {
@@ -181,23 +186,30 @@ exports.getProducts = (req, res, next) => {
         path: '/admin/products'
       });
     })
-    .catch(err => 
-      { const error = new Error(err);
-        error.httpStatusCode =500
-        return next(error);}
-      );
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
-    .catch(err =>{
-      { const error = new Error(err);
-        error.httpStatusCode =500
-        return next(error);}
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
